add bullet list component
